requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
dataclasses-json==0.5.13
distro==1.8.0
dnspython==2.4.0
exceptiongroup==1.1.2
Faker==19.2.0
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.25.1
idna==3.4
importlib-metadata==6.7.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.2
langchain==0.0.335
langchain-experimental==0.0.8
langsmith==0.0.64
loguru==0.7.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==1.2.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
pinecone-client==2.2.2
platformdirs==3.7.0
plotly==5.17.0
prompt-toolkit==3.0.39
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.6.3
PyJWT==2.7.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.11.2
simple-salesforce==1.12.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
SQLAlchemy==2.0.16
tenacity==8.2.2
tqdm==4.65.0
twilio==8.10.1
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
zeep==4.2.1
zipp==3.15.0
